add api and common modules
