+++ backend-rs/src/api/publish.rs
@@ -1,0 +1,3 @@
+pub async fn post_publish() -> &'static str {
+    "Publish"
+}
