+++ backend-rs/src/lib.rs
@@ -1,2 +1,4 @@
+pub mod api;
+pub mod common;
 pub mod search;
 pub mod sql;
